requirements.txt
torch
numpy
tqdm
torchvision
Pillow
tensorboardX